+++ pkg/libwebp/dsp/rescaler_sse2.c.go
@@ -240,7 +240,7 @@ func RescalerExportRowExpand_SSE2(*WebPRescaler const wrk) {
     for (; x_out < x_out_max; ++x_out) {
       const uint32 J = frow[x_out];
       const int v = (int)MULT_FIX(J, wrk.fy_scale);
-      dst[x_out] = (v > 255) ? 255u : (uint8)v;
+      dst[x_out] = (v > 255) ? uint(255) : (uint8)v;
     }
   } else {
     const uint32 B = WEBP_RESCALER_FRAC(-wrk.y_accum, wrk.y_sub);
@@ -272,7 +272,7 @@ func RescalerExportRowExpand_SSE2(*WebPRescaler const wrk) {
       const uint64 I = (uint64)A * frow[x_out] + (uint64)B * irow[x_out];
       const uint32 J = (uint32)((I + ROUNDER) >> WEBP_RESCALER_RFIX);
       const int v = (int)MULT_FIX(J, wrk.fy_scale);
-      dst[x_out] = (v > 255) ? 255u : (uint8)v;
+      dst[x_out] = (v > 255) ? uint(255) : (uint8)v;
     }
   }
 }
@@ -316,7 +316,7 @@ func RescalerExportRowShrink_SSE2(*WebPRescaler const wrk) {
     for (; x_out < x_out_max; ++x_out) {
       const uint32 frac = (int)MULT_FIX_FLOOR(frow[x_out], yscale);
       const int v = (int)MULT_FIX(irow[x_out] - frac, wrk.fxy_scale);
-      dst[x_out] = (v > 255) ? 255u : (uint8)v;
+      dst[x_out] = (v > 255) ? uint(255) : (uint8)v;
       irow[x_out] = frac;  // new fractional start
     }
   } else {
@@ -332,7 +332,7 @@ func RescalerExportRowShrink_SSE2(*WebPRescaler const wrk) {
     }
     for (; x_out < x_out_max; ++x_out) {
       const int v = (int)MULT_FIX(irow[x_out], scale);
-      dst[x_out] = (v > 255) ? 255u : (uint8)v;
+      dst[x_out] = (v > 255) ? uint(255) : (uint8)v;
       irow[x_out] = 0;
     }
   }
