+++ pkg/libwebp/enc/picture_rescale_enc.c.go
@@ -200,7 +200,7 @@ int WebPPictureRescale(*WebPPicture picture, int width, int height) {
   }
 
   if (!picture.use_argb) {
-    work = (rescaler_t*)WebPSafeMalloc(2ULL * width, sizeof(*work));
+    work = (rescaler_t*)WebPSafeMalloc(uint64(2) * width, sizeof(*work));
     if (work == nil) {
       status = VP8_ENC_ERROR_OUT_OF_MEMORY;
       goto Cleanup;
@@ -225,7 +225,7 @@ int WebPPictureRescale(*WebPPicture picture, int width, int height) {
     }
     AlphaMultiplyY(&tmp, 1);
   } else {
-    work = (rescaler_t*)WebPSafeMalloc(2ULL * width * 4, sizeof(*work));
+    work = (rescaler_t*)WebPSafeMalloc(uint64(2) * width * 4, sizeof(*work));
     if (work == nil) {
       status = VP8_ENC_ERROR_BAD_DIMENSION;
       goto Cleanup;
