+++ pkg/libwebp/enc/tree_enc.c.go
@@ -32,7 +32,7 @@ const uint8 VP8CoeffsProba0[NUM_TYPES][NUM_BANDS][NUM_CTX][NUM_PROBAS] = {
 func VP8DefaultProbas(*VP8Encoder const enc) {
   *VP8EncProba const probas = &enc.proba;
   probas.use_skip_proba = 0;
-  memset(probas.segments, 255u, sizeof(probas.segments));
+  memset(probas.segments, uint(255), sizeof(probas.segments));
   memcpy(probas.coeffs, VP8CoeffsProba0, sizeof(VP8CoeffsProba0));
   // Note: we could hard-code the level_costs corresponding to VP8CoeffsProba0, // but that's ~11k of static data. Better call VP8CalculateLevelCosts() later.
   probas.dirty = 1;
